pyproject.toml
[project]
name = "gerry"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["geopandas", "numpy", "pandas", "networkx", "pyyaml", "shapely", "matplotlib"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["gerry"]
